+++ crowdfunding/src/main/kotlin/com/project/indistraw/crowdfunding/application/port/input/dto/RewardDto.kt
@@ -1,12 +1,10 @@
 package com.project.indistraw.crowdfunding.application.port.input.dto
 
-import java.math.BigDecimal
-
 data class RewardDto(
     val idx: Long,
     val title: String,
     val description: String,
-    val price: BigDecimal,
+    val price: Long,
     val imageUrl: String,
     val totalCount: Long?
 )
